Added project Solution/WindowsPhone/SetupLocalIp
The C# part's SetupLocalIp project. This is a tool (console app) used to
setup the host machine's local IP address in files within and without the
project. It's needed for the WinPhone and the Android projects.

For more information, see
* Helpers\Installation Guidelines\Installation Guidelines.docx
* Helpers\Memo\Ha válotzik az IP cím.txt"

TODO: place the tool projects under a common folder, named e.g. Tools